app/pyproject.toml
[project]
name = "func-app-infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-functions>=1.21.0",
    "azure-identity>=1.25.0",
]

[dependency-groups]
dev = [
    "azure-eventhub>=5.15.1",
    "azure-identity>=1.25.3",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "ruff>=0.15.7",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git", "**/__pycache__", "**/.*", "./node_modules"]


